requirements.txt
Django==5.0.1
mysqlclient==2.2.1
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
weasyprint==61.0
pdfkit==1.0.0
django-extensions==3.2.3
whitenoise==6.6.0
django-cors-headers==4.3.1
jdatetime==5.2.0
django-tinymce==4.1.0